backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
PyPDF2
google-generativeai  # if using Gemini
python-decouple
